+++ packages/frontend/src/components/ProjectConnection/UpdateProjectConnection.tsx
@@ -230,7 +230,13 @@ const UpdateProjectConnectionWrapper: FC<{
         return null;
     }
 
-    return <UpdateProjectConnection projectUuid={projectUuid} project={data} />;
+    return (
+        <UpdateProjectConnection
+            key={projectUuid}
+            projectUuid={projectUuid}
+            project={data}
+        />
+    );
 };
 
 export default UpdateProjectConnectionWrapper;